Put decode helpers directly on Request for Response.
The apply method on EntityEncoder is awkward to call when the decoder
takes an implicit parameter.  This gives rise to awkward overloads, like
xml, which surprisingly provide a default without regard to the scoped
implicit.  By moving the call to the request, we can move the implicit
decoder to the end, and remove the syntactic urge for troublesome
overloads.

Request decodings that prefer to be explicit can use the new
`decodeWith` for a reasonable syntax.