Big changes to connection initialization and lifetime management.
__NIO1 and NIO2 now both understand a thread pool__

The resource pool, or NIOXSocketServerGroup, becomes a shared resource. With if you can `.bind(address,service)` and get a ServerChannel. The Server channel can be canceled, but the NIOXSocketServerGroup remains intact.

This makes it simple to easily spin up and down new server services which share a group of threads.

__Change shutdown hook registration behavior__

`addShutdownHook` now returns a Boolean where true signifies successful hook registration.

__Also consolidated examples__