Change the behavior of QuietTimeoutStage
The previous behavior was to timeout on a *pending* request or response after the specified duration. There is a fundemental problem with this: if both a read and a write happen at once and only one succeeds while the other waits, both timeouts are canceled anyway.

The new behavior is probably more intuitive as is.