Like http4s/http4s#332, but with a `Service` arrow.
We bring back `Service` as a specialized Kleisli.  When it's a
type alias, it's easy to mistakenly treat a PartialFunction as
a `Service`.  We want those to get a default, notably as a 404
response when constructing an `HttpService`.  Because Kleisli
operations are common when creating middlewares, this also
cleans some of those up.