ClientTimeoutStage will swallow its own TimeoutException's
Problem: When a client timeouts, the tail stage (the parser) will receive a TimeoutException, as it should. The problem is that the Exception is then pushed out to the head stage making a lot of noise along the way.

Solution: The ClientTimeoutStage should inspect outbound Error commands for a TimeoutException that it has generated and swallow it, sending a Disconnect command instead to sanely shutdown the pipeline.