Optimization of body collection for blaze components
Before this commit the blaze components would see if there wasn't a body and if there is, it would begin a streaming style parsing strategy. This commit changes the second half. The new stragegy examines any remaining bytes to see if that consists of the entire body. If so, wonderful. If not, we go back to the streaming strategy.

This behavior helps the client cleanup more efficiently.