Enable configuration of blaze server parser lengths
Now it is possible to configure the max request line length and max headers length. This
is to avoid denial of service attacks where an evil user sends an endless request line
resulting in OOM on the server.

This behavior was already enabled before but utilized hard coded values for the cutoffs.
Some headers may be huge (think cookie session) so its not out of the question that the
headers could be very large in a minority of situations.