ConnectionManager now says if you're getting a new connection
closes http4s/http4s#571

Previously a closed endpoint would continually spin trying to open a
connection even if one did not exist. This PR addresses that by
modifying the `borrow` function of the `ConnectionManager` to return
a value of type `NextConnection` which includes the boolean parameter
`fresh` describing whether the connection is brand new or not.