Clean up TODO in http2 HeaderDecoder
We need a way to report if the headers are bad but the twitter
implementation doesn't appear to do that. More over, they appear
to be abandoning the library so its unlikely to get fixed. The
only real solution is to find another implementation but everyone
else just bakes their own. :(