Settle on a model
Existential quantification doesn't seem to be as nice of a solution as I'd
hoped so I've settled on a trait with a method which takes a generic argument.
It works great even if the types are not as FP as I'd like. The type signature
is certainly simpler as well.