Clean up some of the concurrency tools
Documentation was incorrect in the actor section. Note: the actors are not
actually used in the library and are just there because its small and might
be useful in the future.

The trampoline executor has been reworked. The old behavior was based on
a simple `ArrayDequeue` which was allocated and destroyed as the executor was
needed and then not needed etc. This is pretty ugly for situations that only
submit a single `Runnable`. The new executor now unrolls the first three
elements of the queue which should cover the vast majority of cases. When
the first three elements spill over, an `ArrayDequeue` is created. After the
queue is emptied, the `ArrayDequeue` is removed to avoid memory leaks.