Add a write loop so we saturate the tcp buffer
We can use a small scratch buffer and avoid copying
a bunch of data that we won't end up using by writing
to the socket in a loop, taking a scratch buffer at
a time and writing it. This may help save resources
when writing extremely large payloads because the
entire payload isn't copied to a direct bytebuffer
just to have most of it not fit into the kernels
socket buffer.