Remove websocket support from raw-blaze
I do this because websockets don't fit into the http world
beyond using http/1.x to upgrade to the websocket protocol.
For services using http2, there is no way to perform this
upgrade and thus you cannot simply use a service which
includes a websocket route transparently with http/1.x and
http/2.

This only effects raw-blaze, and may be reversed in the future,
but as it stands there wasn't a real implementation in this
branch anyway, so lets see how it feels.