Make README.md not scary
The README.md suggests that blaze is unstable and full of bugs when
the reality is that it is reasonably mature and it is the API that
remains in flux as it may need to evolve to support the core http4s
project. This commit updates the README.md to better reflect that
reality.