Use SbtRig for Deployments (#79)
Came with a lot of new scalac options that were updated for as well.

Implicit Widening
Procedure Syntax
Classes/Traits in Package Objects
Warn Value Discard

Were the ones that had significant changes as a result. Would like @rossabaker to make sure the release command looks right and that everything is in place to run a deploy.