Fix a race condition in ServerChannel
When calling ServerChannel.join(), its not clear that tasks may not be
finished executing before the call returns because it's up to the caller
of `close()` to execute all the tasks.

We fix this by making `join()` wait until all tasks have been executed
before returning. To improve the user experience, we also allow passing
an `ExecutionContext` on which to run the shutdown hooks.