Introduce misc tools and the SessionCore
The SessionCore is the most notable part, and is
essentially a glue type that contains references
to the different components of the session. This
allows each component to reach other components
since the SessionCore is often passed as a constructor
parameter. It may not live forever, but is a handy
abstraction for now.