Extract the task queue from the SelectorLoop
Fixes #164

It had a bug where the selector loop may shutdown but a task could be
added to the queue, effectively being lost. Moving the logic out of the
selector loop and into a new type, `TaskQueue`, makes it easier to
reason about and test. Through the new `TaskQueue`s API we can now be
certain that if a task is added to the queue it will be executed.