Fix race condition between shutdown and parsing in Http1ServerStage
The Http1ServerStage doesn't consider whether the stage has been
shutdown from within it's request loop. When the stage is shutdown, it
is reset, which can race with a read-and-parse operation happening in
another thread.

To fix it, we restrict concurrent access to the parser and flag whether
the stage has been shutdown, halting parsing of the prelude if
disconnect has happened.