Remove the `Disconnected` command
As part of the effort to remove the `InboundCommand`s from the Stage
architecture, this remove the `Disconnected` command. The result is that
it is up to the user of the pipeline to free resources and is not
notified eagerly about detected disconnects. In practice, this isn't all
that useful since socket events are similarly lazy: I/O operations are
typically required to detect disconnects.