ServerChannelGroup.acceptConnection is moved to `BufferedPipelinebuilder` (#191)
Closes #184.

The acceptConnection method is in a silly place in that the
`ServerSocketGroup` represents a collection of resources that are
capible of listening for socket connections. It is the wrong place to be
defining the logic for accepting or rejecting connections. This is
better done via changing the definition of `BufferPipelineBuilder` from
`SocketConnection => LeafBuilder[ByteBuffer]`
to
`SocketConnection => Future[LeafBuilder[ByteBuffer]]`
It also makes sense to rename it to `SocketPipelineBuilder` since it
must be dealing in `SocketConnection`s.