Tails are in charge of closing resources
Right now, when a channel fails to read or write it fails the channel,
closing everything etc. This makes it impossible to handle half-closed
scenarios. Instead of the Head being in charge of releasing resources,
we move that responsibility to the tails. This allows them to interact
with a half closed socket.

Note: this doesn't yet allow half-closing a socket, though it is a
necessary step.