NIO1HeadStage can now read without waiting on NIO notification
Currently, reading and writing requires notification via NIO interests.
This can be really wasteful if there is already data in the inbound buffer
or space in the outbound buffer.

Now we try to read and write data initially, and only if we can't do we
fall back to interest operations.