Add ability to close a BodyWriter with an exception
Right now, there is no way for a user of a BodyWriter to signal to the
consuming side that it is being closed due to some problem, thus the
other end may think that the stream closed successfully, while in
reality it may be incomplete.

To fix this, we add a `cause: Option[Throwable]` to the `close` method
to allow the data producer to signal that it is closing due to a
problem.